fixed edit order. Added EditOrderTests with Effort

+++ Northwind.DataLayer/Infrastructure/IRepository.cs
@@ -11,7 +11,7 @@ public interface IRepository<T> where T : class
     void Add(T entity);
     void Delete(T entity);
     void Delete(Expression<Func<T, bool>> where);
-    T GetById(long Id);
+    T GetById(int Id);
     T GetById(string Id);
     T Get(Expression<Func<T, bool>> where);
     IEnumerable<T> GetAll();
